python/requirements.txt
pyvista
pyvistaqt
vtk
PyQt5
numpy
scipy
matplotlib
appdirs
joblib
Pillow
